house-scorecard-backend/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.2
sentry-sdk[django]==2.17.0
gunicorn==23.0.0
dj-database-url==2.2.0
whitenoise==6.8.2
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0

# AI Analysis Dependencies
python-dotenv==1.0.1
google-genai==0.9.0
pillow==11.1.0

# Background Task Processing
celery==5.4.0
django-celery-results==2.5.1
redis==5.2.0
kombu==5.4.2

# Additional utilities for production
python-dateutil==2.9.0
